[GLX] Implement __glXFetchDispatchEntry()
This function is an implementation of the fetchDispatchEntry export,
which is called by vendor-implemented GLX dipsatchers to fetch a
dispatch table entry given a logical dispatch table index.

This change also implements the logic for assigning dispatch table
indices to GLX entrypoints at glXGetProcAddress() time.  The dynamic
dispatch table is actually a hashtable which uses the "index" as a key,
and not a flat array, to prevent wasted memory usage from sparse
dispatch tables.  Since GLX entrypoints are generally not
performance-critical this should be fine.