Add libGL and libOpenGL filter libraries
libOpenGL is a filter library on libGLdispatch which only exports the
core GL 1.x-4.x entrypoints.

libGL is a filter library on libGLX and libGLdispatch which only exports
the core GL 1.x-4.x entrypoints and GLX 1.4 entrypoints. This is a
backwards compatibility layer for legacy apps which link against the old
libGL.so ABI.