[mapi] Implement glapi_init_table_from_callback()
This is a new glapi function which takes an already-allocated
_glapi_table and fills it in using the given getProcAddress callback.
This is needed because the existing _glapi_create_table_from_handle()
function only allows clients to pass in a loader handle from which to
dlsym() entrypoints. This reduces the number of symbols required by
each vendor library and also allows vendor libraries to provide
different context-dependent entrypoints if desired when building
dispatch tables.