[build] enable silent rules by default
Automake still defaults to verbose output unless you explicitly request that the
silent rules be enabled by default.  It's possible to only enable the silent
rules on versions of automake that actually support it, but automake 1.11 has
been around for long enough that it seems easier to just bump the minimum
required version.

Add the boilerplate required to turn on the silent rules by default.  They can
still be disabled by passing --disable-silent-rules to configure or by passing
V=0 on the make command line.

Signed-off-by: Aaron Plattner <aplattner@nvidia.com>