Disable aliasing static dispatch functions
The aliasing functionality designed to reduce the size of the static
dispatch table in glapi interacts badly with forward-compatible contexts
as defined in GLX_ARB_create_context.  For example, a forward-compatible
OpenGL ES 1.0 context will not support calling into glGetTexGeniv(), but
may support calling into glGetTexGenivOES().  These two functions are aliases
in glapi, so a call to the latter may end up dispatching into the former,
which may be implemented as a stub in the vendor library.

Hence, when generating the static dispatch functions via gl_XML.py,
don't use the "alias" attribute if it is present in the "function" tag.

Signed-off-by: Brian Nguyen <brnguyen@nvidia.com>