Fix glXGetConfig() error return codes
The spec mandates that glXGetConfig() should return GLX_BAD_VALUE if
any parameter is invalid, and GLX_BAD_SCREEN if the screen of the
<vis> parameter does not correspond to a screen.

Hence, check that the arguments of glXGetConfig() are non-NULL and
return GLX_BAD_VALUE if otherwise.  Furthermore, update the noop
stub to return GLX_BAD_SCREEN, since the vendor-neutral dispatcher
will only choose the no-op stub if it couldn't find a valid vendor
for the given screen number.

Signed-off-by: Brian Nguyen <brnguyen@nvidia.com>