Don't throw X errors if an x11glvnd request fails
Failing to lookup a mapping via this extension should not cause GLX
to throw X errors.  Hence, instead of returning an error code in
the server-side functions if the lookup fails, write a reply which
indicates the lookup failed.

With this change, ProcGLVQueryXIDScreenMapping() sends a reply with
screen -1 if its lookup failed, and ProcGLVQueryScreenVendorMapping()
sends a reply with a 0-length vendor string if its lookup failed.

Signed-off-by: Brian Nguyen <brnguyen@nvidia.com>