Add XCloseDisplay() hook to x11glvnd, and allow clients to register their own hooks
Use XEXT_GENERATE_CLOSE_DISPLAY() to generate an XCloseDisplay() hook
close_display_internal(); this is necessary for the x11glvnd extension
to clean up its private data structure when the display is closed.  Wrap
this hook with a separate close_display() hook; this hook calls
functions in a callback list.

Add XGLVRegisterCloseDisplayCallback() to allow x11glvnd clients to
register these XCloseDisplay() callbacks with x11glvnd; this allows
clients to hook into XCloseDisplay() without needing to implement
X11 extension shim code.

In libGLX, use XGLVRegisterCloseDisplayCallback() to register a callback
to remove references to a closed display in the API state hashtable.

Bonus: add a missing SyncHandle() in the CHECK_EXTENSION() macro.

Signed-off-by: Brian Nguyen <brnguyen@nvidia.com>