[libglx] Update ABI to notify vendors of X errors, fix spec compliance issue
This commit adds a new vendor callback, notifyError(), for notifying a
vendor library when an X error should be generated.  This allows the
vendor to perform any needed error recovery and to generate the error.

In MakeContextCurrentInternal(), check for:

    (context && (draw == None || read == None)) ||
    (!context && (draw != None || read != None))

and call notifyError() on an appropriate vendor to generate BadMatch
before returning False.

Signed-off-by: Brian Nguyen <brnguyen@nvidia.com>