[libglx] Relax thread checks in MakeCurrent
Top-level dispatch here can't do this check, because it would break
GLX_MESA_multithread_makecurrent; and the check as it is doesn't raise
BadAccess like it should, so it's broken anyway.  Instead just pass the
request through to the server (like we have to anyway) and let the
server raise the error if needed.

Signed-off-by: Adam Jackson <ajax@redhat.com>