libglvnd: Add XGLVQueryExtension, fixes for XGLVQueryVersion.
Added a XGLVQueryExtension function, and changed XGLVQueryVersion to generate
an error if the extension isn't supported by the server, so that they match
the behavior of other X extensions.

Fixed the padding in the reply structures so that they're all 32 bytes long.

Fixed the order of the request handler array in the server.