libGLX: Fixes for when x11glvnd isn't available.
If the server doesn't support x11glvnd, then use the fallback vendor name
instead of trying to ask the server for a name.

Fix __glXVendorFromDrawable to work without x11glvnd. Since every screen will
use the same vendor in that case, it just returns the vendor from screen zero.

Fix __glXGetDrawableStaticDispatch. It now uses __glXVendorFromDrawable to look
up the vendor for a drawable.