libGLdispatch: Work on adding x86-64 TSD stubs.
Added a new file entry_x86-64_tsd.h. It's just a copy of entry_x86_tsd.h for
now, since that's the interface that it has to match.

Updated entry.c to use the new file depending on the GLX_USE_TLS macro, just
like it does for x86.

Added a __GLDISPATCH_STUB_X86_64_TSD enum to GLdispatchABI.h, and a
matching ENTRY_X86_64_TSD enum to entry.h.

Added a "--disable-tls" flag to the configure script to tell it not to use a
TLS version.