libGLdispatch: Finish x86-64 TSD stubs.
Implement the assembly code for the x86-64 TSD stubs.

In u_execmem.c, allocate enough space for 64-byte stubs, which is the size of
the new x86-64 TSD stubs.

The new stubs don't fit in 32 bytes, because they have to hold two movabs
instructions to deal with 64-bit addresses, plus push and pop instructions to
preserve the six registers that are used for passing arguments.