Add a new functions for allocating executable memory.
Added two new functions, AllocExecPages and FreeExecPages, for allocating
executable memory for generating stub functions.

It uses a single mmap right now, but it's set up to return two pointers, so it
will be able to return separate writable and executable mappings. That will let
it work on systems that don't allow pages to be both writable and executable.

The new functions are in utils_misc.c, since it will be needed in both libGLX
and libGLXdispatch.