libGLX: Clean up MakeCurrent.
Moved the logic in glXMakeCurrent and glXMakeContextCurrent to a common
function, CommonMakeCurrent.

glXMakeCurrent/glXMakeContextCurrent will now return early if the old and
new context and drawables are the same.

UpdateCurrentContext no longer tries to look up the current context from TSD.
The MakeCurrent functions already look up the current context, so they just
pass the old context to UpdateCurrentContext.