Remove the TSD key from libGLX.
Added a pointer to a callback function __GLdispatchAPIState to notify the
window system library of thread termination.

Removed tsdContextKey from libglx.c. It now uses the API state stored in
libGLdispatch and the callback function in __GLdispatchAPIState instead.

libGLX will now free the __GLXAPIState struct for a thread when the thread
terminates, instead of waiting until __glXAPITeardown.

Fixes issue #35