libGLX: Clean up MakeCurrent handling.
Change the code for glXMakeCurrent/glXMakeCurrentContext to better handle
different changes.

If it switches between two contexts that use the same vendor, it will not call
into libGLdispatch, since it would just be switching to all the same
parameters.

If it switches between two contexts that use different vendors, then it will
fully release the current context before making the new context current.

This makes it easier to ensure that libGLX and the vendor libraries stay in a
consistent state, even when recovering from a failure in the vendor library or
in libGLdispatch.

This also means that __glDispatchMakeCurrent is only called when the current
thread has no current API context, so libGLdispatch doesn't need to deal with
switching between two different API states.