libGLX: Simplify UpdateCurrentContext.
UpdateCurrentContext will now re-use the existing hashtable entry if it's
switching between two contexts, so there's no longer a possibility of malloc
failing. The only case where it can fail now is if there is a new context but
not an old context.

UpdateCurrentContext will handle removing the old context from the context to
screen mapping as appropriate.

In CommonMakeCurrent, when releasing a context or switching between two
contexts with the same vendor, it will now call UpdateCurrentContext once at
the end, since UpdateCurrentContext can't fail in those cases.