libGLdispatch: Add an ABI version number.
Add a version number for the interface between libGLX and libGLdispatch, which
libGLX checks during initialization.

The interface between libGLX and libGLdispatch can and should be a stable ABI,
so that it's more resiliant to version mismatches if they come from different
packages or installers.

The version number is mainly a sanity check in case some future version of
libGLdispatch does change the interface.