libGLdispatch: Use separate write and execute mappings.
Change libGLdispatch so that it will use separate writable and executable
mappings when it generates dispatch stubs.

Added a function, u_execmem_get_writable that will hand back a writable pointer
for an executable pointer. That avoids having to rework everything else in
libGLdispatch to keep track of both pointers.

The static x86-64 dispatch stubs are now marked as read+execute instead of
read+write+execute. Before patching them, it will call mprotect to make them
writable.

Change __GLdispatchStubPatchCallbacks::restoreFuncs so that it returns a flag
to indicate success or failure.

Issue #37