libGLX: Remove the isClientAPI flag.
Remove the isClientAPI parameter from __GLXvendorCallbacksRec::getProcAddress.

The only thing that libGLX can use to determine the the isClientAPI flag is the
function name, which a vendor library could implement just as easily. In
addition, since it would be possible for a GL entrypoint to start with "glX",
the isClientAPI flag isn't always reliable.