libGLdispatch: Clean up the entry code.
Removed some unused files and functions.

Removed _GLAPI_EXPORT on a bunch of functions that don't need to be exported.

Removed some unnecessary macros:

Removed the macro u_current, and replaced all references to it with
_glapi_Current or _glapi_tls_Current.

Removed the mapi_table macro, and replaced it with _glapi_table.

Removed the ENTRY_CURRENT_TABLE and ENTRY_CURRENT_TABLE_GET macros.

Cleaned up the code related to storing the current dispatch table:

Rearranged things so that _glapi_* variables and functions are the public
interface to everything else, and the u_current_* functions are the internal
implementation.

Changed _glapi_Current to use the same definition on all builds.

Removed the existing u_current_get and replaced it with u_current_get_internal,
to match u_current_set.

Made _glapi_get_dispatch into a separate (exported) function, rather than a
macro for u_current_get_internal.

Consolidated most of the code for keeping track of the current dispatch table
into u_current.c. Rearranged the file to use a single #if/#else block to
separate the TLS and TSD versions.

Removed all of the non-pthread code. There's no reason to keep it around, since
the rest of libglvnd won't build without pthreads.

Removed most of the contents of u_thread.h, and replaced it with direct calls
to the pthreads functions.