tests: Fix getProcAddress for unknown functions.
Change the dummy vendor library's getProcAddress callback so that it will
return NULL for an unknown function, not a NOP stub.

Everything in libGLX should be prepared to deal with getting a NULL from the
vendor library, and libGLX even depends on that to distinguish between GL and
GLX functions.