libGLX: Don't leak __GLXAPIState structs.
libGLX only gets a callback at thread termination if it has a current context.
If there isn't a current context, then it will leave the thread's __GLXAPIState
in the hashtable until teardown.

Now, libGLX will only keep a __GLXAPIState when it has a current
context. When glXMakeCurrent is called to release the current context, it will
free the __GLXAPIState.

Removed the hashtable for keeping track of the __GLXAPIStates and replaced it
with a linked list.