libGLX: Change ABI to support glXGetContextIDEXT.
Changed __GLXapiExports::vendorFromContext so that it only takes a GLXContext
parameter as input, and returns the display, screen, and vendor for the
context.

This is needed to support a dispatch function for glXGetContextIDEXT, since it
only takes a GLXContext handle, not a display.

In libglxmapping.c, the __GLXscreenPointerMappingHash struct now keeps the
display pointer along with the screen number.

Also added an error check for GLXFBConfigs to make sure that the display passed
from the caller matches the recorded one.