libGLdispatch.so: Fixes for the public stub array.
Fix a couple of places that assume that the public stub array (as defined in
entry*.c) is in the same order as the dispatch table.

stub_get_addr will now calculate the index of the stub instead of assuming that
the slot and index are the same.

In stubRestoreFuncsInternal, it will run through the length of the public_stubs
array instead of assuming that the array is of length MAPI_TABLE_NUM_STATIC.

This will allow us to remove all the entrypoints in libOpenGL.so that we're
not going to export.