libOpenGL.so: Only generate the entrypoints that we're going to export.
gen_gldispatch_mapi.py now takes an extra parameter to specify which library
it's building, so that it can generate the correct function list for that
library. For libOpenGL.so, it will only generate the functions that it's going
to export.

Updated the makefiles for libOpenGL.so, libGL.so, and libGLdispatch.so to
generate the correct set of entrypoints for each library.

Moved the code to figure out the libOpenGL.so function names to genCommon.py.