libGLdispatch: Enable entrypoint patching for the x86 TSD build.
Rewrote the x86 TSD stubs to follow the same pattern and use the same common
functions as the x86-64 stubs.

Aside from being shorter and simpler with less code duplicaiton, this will
allow entrypoint patching to work in the x86 build.

Fixed some errors in the entrypoint patching in the unit tests. The
libGLX_dummy.so library was patching in PC-relative addresses for the MOV
instructions, but those instructions expect absolute addresses.