GLX: Add support for GLX_MESA_multithread_makecurrent.
Rework the context tracking in libGLX.so so that it can handle a context that's
current to more than one thread at a time.

There's now a single hashtable to keep track of contexts, which replaces both
the (context -> vendor) map and the current context map.

The new __GLXcontextInfo struct keeps track of the number of threads where the
context is current, and whether the context is marked for deletion.

This also provides a workaround of sorts for some broken applications, which
try (and fail) to make a context current on two threads at the same time, and
aren't prepared to deal with the resulting BadAccess error. Moving the
duplicate context check into the vendor library allows it to preserve that
pre-libglvnd behavior for apps that depend on it.

Reviewed-by: Andy Ritger <aritger@nvidia.com>