GLX: Cleanup vendor library loading.
The __GLXvendorNameHash, __GLXvendorInfo, and vendor name are now allocated in
a single malloc.

Remove some duplicated cleanup code between TeardownVendor and
__glXLookupVendorByName.

Merge TeardownVendor into CleanupVendorNameInto.

In __glXLookupVendorByName, simplify the cleanup path on error. Remove the
local variables that are just copies of the __GLXvendorInfo members.

Remove InitExportsTable and statically initialize the glxExportsTable struct.

Reviewed-by: Andy Ritger <aritger@nvidia.com>