uthash: update to latest 044474afe5a57db4302cc722de476d20d1ab70c7
Since out initial import the project has gained a lot of fixes - missing
brackets around arguments in macros, missing error handling and others.

The project owner has even announced that there won't be any work on it.

Note: this commit introduces over 7000 of whitespace changes, so use the
"-w" option of git diff/show and others to minimise the noise.

Signed-off-by:    Emil Velikov <emil.velikov@collabora.com>