Remove auxiliary ELF symbol filtering
The code was disabled due to a glibc bug 16272. which as of this point
remains open.

As n alternative solution is already in place and fix for the mentioned
issue is unlikely to land in order distributions, it does not make it a
viable solution that we can use (in the near future).

Drop the code for now, to prevent anyone building a completely broken
library.

Signed-off-by: Emil Velikov <emil.velikov@collabora.com>