GLX: Allow different dispatch tables for different contexts.
Added two new optional callbacks to __GLXapiImports, which allow a vendor
library to use different dispatch tables for different rendering contexts.

The vendor library can provide an opaque dispatch table handle for each
context, which libGLX.so will use to assign and populate a dispatch table to
use when that context is current.

Incremented the ABI version number for libGLX to 1.1.